No need to change docker configuration anymore
We now simply say to dnsmasq to listen Docker0 interface and we say to
resolved to listen this DNS server. Docker is configured to use
resolved, so the domains are directly accessibles.

This way, if the configuration is reverted, docker will use the default
DNS server.

That's better now, we don't need to know the docker0 address to inject
in `/etc/docker/daemon.json` anymore.